ai-engine-python/requirements.txt
pymongo
pandas
scikit-learn
python-dotenv
paho-mqtt
fastapi
uvicorn